Part 1 finished

+++ src/views/PIPCalcTextView.java
@@ -87,7 +87,7 @@ public class PIPCalcTextView implements Observer{
      */
     @Override
     public void update(Observable o, Object arg) {
-        System.out.println((Integer)arg);
+        System.out.println(arg);
     }
 
     /**
